refactor(api/policy): split registerPolicyRoutes into per-entity registrars (#225)
Second slice of #225. `registerPolicyRoutes` (~1330 lines) becomes a
75-line composition root that calls one focused registrar per entity, in
the original registration order:

- New `api/policy/routes/{users,clients,links,activities,activity-groups,
  user-groups,group-schedules,group-exceptions,group-budgets,budgets,
  schedules,exceptions,notification-policy}.ts`, each registering its own
  5–7 endpoints and reusing the Phase-1 shared helpers/builders.
- Each registrar takes `(scope, push)` (activities / activity-groups /
  user-groups take only `scope` — grouping definitions has no per-client
  push effect) and derives its own `typed`/`guard` locally, so the files
  are self-contained.

Behaviour-preserving: no route path, status code, error envelope, push
fan-out, or validation changed. All existing `tests/api/*` route suites
stay green unchanged; `api/policy/routes` sits at ~96.8% coverage.
